BaseTools: Update TE base address even if it has no relocs
It is possible that all the relocations in a TE section are relative
and need no processing. In this case no Base relocs may exist and the
code can run at a rebased location. When this TE Image is rebased we
still need to update the Base Address for the runtime loader to behave
correctly.

Signed-off-by: Harry Liebel <Harry.Liebel@arm.com>

Change-Id: I990442d4ec3c177b6a618565b04a77467e6acb2e