EmbeddedPkg/FdtPlatformDxe: Run FDT installation process at TPL_APPLICATION level
Run FDT installation process at the end of the DXE phase at TPL_APPLICATION
level and not TPL_CALLBACK level.

The FDT installation process was run when the "end of DXE" event was signalled
and thus was run at TPL_CALLBACK level. This caused the boot process to be
stuck in UEFI when the FDT had to be fetched from a TFTP server as the IP
stack rely on timers whose callback function is called at TPL_CALLBACK level.

This patch adds the installation of the INSTALL_FDT_PROTOCOL in the FDT platform
driver. This protocol is then used at the beginning of the BDS to launch
the FDT installation process at TPL_APPLICATION level.

Change-Id: Ib711231bdaed55ce4dcbc7026427e7283fcfc986