EmbeddedPkg: Added Marvell Yukon Ethernet support
This change adds support for the PCI Marvell Yukon Ethernet
controller.
This ethernet controller can be found on the ARM Juno development
platform.
The UEFI driver has been tested on early development platform and
was known to work. But it has not been tested on the ARM Juno
Development platform.

Change-Id: Ifb43451d2acd1e0c42c3bc55c08c23af45e6780e
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Dan Handley <Dan.Handley@arm.com>
Signed-off-by: Olivier Martin <Olivier.Martin@arm.com>